# specs/04_api_spec.yaml
openapi: 3.0.0
info:
  title: MTF Trading API
  version: 0.1.0
paths:
...
  /api/v1/deployments/:
    get:
      summary: List all deployments
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_DeploymentList'
    post:
      summary: Create a new deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentCreate'
      responses:
        '200':
          description: Created deployment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_Deployment'
  /api/v1/deployments/{id}/stop:
    post:
      summary: Stop a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deployment stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse_Deployment'
...
components:
  schemas:
...
    ResponseStatus:
      type: string
      enum: [success, error, fail]
    ErrorCode:
      type: string
      enum:
        - AUTH_1001
        - AUTH_1002
        - AUTH_1003
        - AUTH_1004
        - AUTH_1005
        - VAL_2001
        - VAL_2002
        - VAL_2003
        - VAL_2004
        - VAL_2005
        - VAL_2006
        - VAL_2007
        - RES_3001
        - RES_3002
        - RES_3003
        - BIZ_4001
        - BIZ_4002
        - BIZ_4003
        - BIZ_4004
        - RATE_5001
        - RATE_5002
        - SRV_9001
        - SRV_9002
        - SRV_9003
        - SRV_9004
...
    APIResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          $ref: '#/components/schemas/ResponseStatus'
        data:
          type: object
          nullable: true
        message:
          type: string
          nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        meta:
          $ref: '#/components/schemas/Meta'
        auth:
          $ref: '#/components/schemas/AuthTokens'
        rate_limit:
          $ref: '#/components/schemas/RateLimitInfo'
        timestamp:
          type: string
          format: date-time
...
    Meta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    AuthTokens:
      type: object
      required:
        - access_token
        - expires_in
        - expires_at
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
        expires_at:
          type: string
          format: date-time
    RateLimitInfo:
      type: object
      required:
        - limit
        - remaining
        - reset
        - reset_in_seconds
      properties:
        limit:
          type: integer
        remaining:
          type: integer
        reset:
          type: string
          format: date-time
        reset_in_seconds:
          type: integer
    ErrorDetail:
      type: object
      required:
        - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          $ref: '#/components/schemas/ErrorCode'

    # --- Generic Wrappers (instantiated for each type) ---
    # Note: OpenAPI 3.0 doesn't support generics directly, so we define explicit types.
...
    Deployment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        strategy_id:
          type: string
          format: uuid
        stock_symbol:
          type: string
        timeframe:
          type: string
        status:
          type: string
          enum: [STARTING, ACTIVE, STOPPED, ERROR, STOPPING]
        is_live:
          type: boolean
        config_snapshot:
          type: object
        last_error:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
        stopped_at:
          type: string
          format: date-time
          nullable: true
    DeploymentCreate:
      type: object
      required:
        - strategy_id
        - stock_symbol
        - timeframe
        - config_snapshot
      properties:
        strategy_id:
          type: string
          format: uuid
        stock_symbol:
          type: string
        timeframe:
          type: string
        is_live:
          type: boolean
        config_snapshot:
          type: object
    APIResponse_DeploymentList:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Deployment'
    APIResponse_Deployment:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Deployment'

    # --- AI Chat Schemas ---
...